requirements.txt
# TESP API Python Dependencies
# Python 3.10.0 or higher required

# Core dependencies
aio_pika>=9.5.7,<10.0.0
fastapi>=0.75.1,<0.76.0
orjson>=3.6.8,<4.0.0
gunicorn>=20.1.0,<21.0.0
uvicorn>=0.17.6,<0.18.0
pydantic>=1.9.0,<2.0.0
dynaconf>=3.1.8,<4.0.0
motor>=3.0.0,<4.0.0
loguru>=0.6.0,<1.0.0
aiohttp>=3.13.3,<4.0.0
aioftp>=0.21.0,<1.0.0
PyMonad>=2.4.0,<3.0.0
aiobotocore>=2.4.2,<3.0.0
requests>=2.28.2,<3.0.0

# OAuth2/JWT dependencies with crypto extras
pyjwt>=2.8.0,<3.0.0
cryptography>=41.0.5,<42.0.0

# Development dependencies
pytest>=7.1.1,<8.0.0
